Cambiado usando long

+++ com.entornos.NicolasSimpsonPadilla.algoritmos/src/com/entornos/NicolasSimpsonPadilla/algoritmos/algoritmos.java
@@ -3,9 +3,9 @@ package com.entornos.NicolasSimpsonPadilla.algoritmos;
 public class algoritmos {
 
 
-	public static int fibonacci(int n) {
-		int fibo1=1;
-		int fibo2=1;
+	public static long fibonacci(int n) {
+		long fibo1=1;
+		long fibo2=1;
 		for (int i=2;i<=n;i++) {
 			fibo2 = fibo1 + fibo2;
             fibo1 = fibo2 - fibo1;
@@ -13,8 +13,8 @@ public class algoritmos {
 		return fibo2;
 	}
 	
-	public static int factorial (int n) {
-		int factorial = 1;
+	public static long factorial (int n) {
+		long factorial = 1;
 		for (int i = 1; i <= n; i++) {
             factorial = factorial * i;
         }
